// src/app/model/SearchItemResultRow.ts
export class SearchItemResultRow {
    callNumber: string;
    chronologyAndEnum: string;
    customerCode: string;
    barcode: string;
    useRestriction: string;
    collectionGroupDesignation: string;
    availability: string;
    selectedItem: boolean;
    itemId: number;
    owningInstitutionItemId: string;
    owningInstitutionHoldingsId: string;
}